Move IWebRequestHandlerJson Fixes #49
