+++ src/types/ITriggerJson.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 import IMqttConfigJson from '../handlers/mqttManager/IMqttConfigJson';
-import IWebRequestHandlerJson from './IWebRequestHandlerJson';
+import IWebRequestHandlerJson from '../handlers/webRequest/IWebRequestHandlerJson';
 
 export default interface ITriggerJson {
   cooldownTime: number;
